+++ extensions/Splitwise/src/hooks/userPreferences.tsx
@@ -1,10 +1,3 @@
 import { getPreferenceValues } from "@raycast/api";
 
-const { personalAccessToken } = getPreferenceValues();
 export const { loadingLimit } = getPreferenceValues();
-
-export const HEADER = {
-  headers: {
-    Authorization: `Bearer ${personalAccessToken}`,
-  },
-};